Fixed Issue #665 by adding support for Windows
Added support for Windows having preexisting files when using fget_object

+++ minio/api.py
@@ -599,7 +599,13 @@ class Minio(object):
                 raise InvalidSizeError(msg)
 
         # Rename with destination file.
-        os.rename(file_part_path, file_path)
+        if !os.path.exists(file_path)
+             os.rename(file_part_path, file_path)
+        else:
+            #If the file already exists, need to remove it for Windows OS compatibility
+            os.remove(file_path)
+            os.rename(file_part_path, file_path)
+        
 
         # Return the stat
         return stat
